Fix Win64 build problem introduced by r5041.
Review URL: http://codereview.chromium.org/2893009

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@5042 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/heap.cc
@@ -818,7 +818,7 @@ class PromotionQueue {
 
   void remove(HeapObject** target, int* size) {
     *target = reinterpret_cast<HeapObject*>(*(--front_));
-    *size = *(--front_);
+    *size = static_cast<int>(*(--front_));
     // Assert no underflow.
     ASSERT(front_ >= rear_);
   }
